requirements.txt
pip~=23.2.1
Django~=5.1.3
attrs~=24.2.0
Pillow~=9.1.1
asgiref~=3.8.1
sqlparse~=0.5.2
psycopg~=3.1.12
psycopg2-binary~=2.9.10
PyYAML~=6.0.2
djangorestframework~=3.13.1